CRM System - starting workspace
The workspace has working spring mvc CRM application. Rest concept will
be built over this project.
Build the project with ant script and it will be deployed in tomcat
folder which should be in the same folder with the CRM System project
else change the path in script.
Welcome page - http://localhost:8080/mywebapp/all-customers.html